account-exception-app/src/com/monocept/model/Account.java
package com.monocept.model;

public class Account {
	private final int accountNumber;
	private final String name;
	private double balance;

	public Account(int accountNumber, String name, double balance) {
		this.accountNumber = accountNumber;
		this.name = name;
		this.balance = balance;
	}

	public Account(int accountNumber, String name) {
		this(accountNumber, name, 500);
	}

	public void deposit(double depositAmount) {
		this.balance += depositAmount;
	}

	public void withdraw(double withdrawAmount) throws Exception {
		if (this.balance - withdrawAmount < 500) {
			throw new Exception("cannot withdraw " + withdrawAmount + ",min-balance limit 500");
		}
		this.balance -= withdrawAmount;
	}

	public int getAccountNumber() {
		return accountNumber;
	}

	public String getName() {
		return name;
	}

	public double getBalance() {
		return balance;
	}

}
